+++ internal/providergen/ir.ts
@@ -1,3 +1,4 @@
+import { camelize, humanize, pluralize, singularize } from "inflection";
 import { match, P } from "ts-pattern";
 
 interface IRBase {
@@ -112,4 +113,83 @@ export function toIR({
         `Unsupported swagger schema type: ${JSON.stringify(schema)}`
       );
     });
+}
+
+export function generateResourceIR({
+  swagger,
+  name,
+}: {
+  swagger: any;
+  name: string;
+}) {
+  const resourceSchema = swagger.components.schemas[name];
+  if (!resourceSchema) {
+    throw new Error(`Resource ${name} not found`);
+  }
+
+  const newResourceSchema = swagger.components.schemas[`new_${name}`];
+  if (!newResourceSchema) {
+    throw new Error(`New resource ${name} not found`);
+  }
+
+  const collectionSchema = Object.entries(
+    swagger.paths as Record<string, any>
+  ).find(
+    ([_, pathSchema]) =>
+      pathSchema.get &&
+      pathSchema.get.operationId === `list${camelize(pluralize(name))}`
+  )?.[1];
+  if (!collectionSchema) {
+    throw new Error(`List path for ${name} not found`);
+  }
+
+  const getSchema = Object.entries(swagger.paths as Record<string, any>).find(
+    ([_, pathSchema]) =>
+      pathSchema.get &&
+      pathSchema.get.operationId === `get${camelize(singularize(name))}`
+  )?.[1]?.get;
+  if (!getSchema) {
+    throw new Error(`Get path for ${name} not found`);
+  }
+
+  // Get path ID parameter
+  const pathIdParameter = collectionSchema?.parameters?.[0]?.name as
+    | string
+    | undefined;
+  const pathIdIR = pathIdParameter
+    ? toIR({
+        schema: resourceSchema.properties[pathIdParameter],
+        required: null,
+      })
+    : null;
+
+  const getHasQueryParams =
+    getSchema?.parameters?.some((param: any) => param.in === "query") ?? false;
+
+  // Generate immediate representation of the resource
+  const irFields = toIR({
+    schema: resourceSchema,
+    required: newResourceSchema.required,
+  });
+  if (irFields.kind !== "object") {
+    throw new Error("Resource root must be an object");
+  }
+
+  const ir: IRResource = {
+    kind: "resource",
+    resourceType: name,
+    listPathIdParam:
+      pathIdParameter && pathIdIR
+        ? { name: pathIdParameter, element: pathIdIR }
+        : null,
+    getHasQueryParams,
+    idElement: {
+      kind: "string",
+      computedOptionalRequired: "computed",
+      description: `The ID of the ${humanize(name, true)}`,
+    },
+    fields: irFields.fields,
+  };
+
+  return ir;
 }
